requirements.txt
web3~=7.6.0
termcolor~=2.5.0
python-dotenv~=1.0.1